UI: First pass at creating some tooltip descriptions for the various editor types
While some of these are rather rough and crude, they should be sufficient to provide
some helpful hints to (new) users when mousing over the editor type selectors
(i.e. now we can see the names of the editors; without the descriptions, the enum
item names wouldn't get shown) and also when browsing the list.

This is useful now that we have the ability to display tooltips in menus, thanks
to Campbell's earlier commit (35f62bd)